// src/app/models/proyecto.models.ts
export interface Project {
  id: string;
  name: string;
  id_categoria: string;
  imagen: string;
  imagen_principal: string;
  imagen_secundaria: string;
  banner: string;
  descripcion: string;
  ubicacion: string;
  area_terreno: string;
  area_construida: string;
  numero_pisos: string;
  fotos: string[];
}